Move country context and experiment tracking out of effects
